first five boxes on issue #35
