+++ test/polyfills.js
@@ -97,7 +97,7 @@ export const polyfill = () => {
                     const [m11, m12, m13, m14, m21, m22, m23, m24, m31, m32,
                         m33, m34, m41, m42, m43, m44] = array;
 
-                    this.#is2D = true;
+                    this.#is2D = false;
 
                     this.#m11 = m11;
                     this.m12 = m12;
