+++ test/test.js
@@ -44,6 +44,30 @@ try {
     // There are so many warnings logged that it slows the test down
     console.warn = () => {};
 
+    describe('index.js', () => {
+        test('exports cloneDeep as a default export. Also exports CLONE ' +
+             'symbol, cloneDeepFully, and useCustomizers as named exports', 
+             async () => {
+            const module = await import('../index.js');
+            const array = [];
+            const getPusher = n => () => array.push(n);
+
+            const clone = module['default']({ a: 'a' });
+            const fullClone = module['cloneDeepFully']({ a: 'a' });
+            module['useCustomizers']([
+                getPusher(1), 
+                getPusher(2)
+            ])();
+            const symbol = module['CLONE'];
+
+            assert.strictEqual(4, Object.keys(module).length);
+            assert.deepEqual(clone, { a: 'a' });
+            assert.deepEqual(fullClone, { a: 'a' });
+            assert.deepEqual(array, [1, 2]);
+            assert.strictEqual('symbol', typeof symbol);
+        });
+    });
+
     describe('cloneDeep without customizer', () => {
 
         const number = 1234566;
